objectify-bluetooth-hardware-requests Fixed preformence on requests, renamed some stuff, general improvements.
