+++ BetterJoy/Hardware/SubCommand.cs
@@ -1,0 +1,107 @@
+#nullable enable
+using System;
+using System.Runtime.CompilerServices;
+using System.Text;
+
+namespace BetterJoy.Hardware;
+
+public class SubCommand
+{
+    private static readonly byte[] _stopRumbleBuf = [0x0, 0x1, 0x40, 0x40, 0x0, 0x1, 0x40, 0x40]; // Stop rumble
+    
+    private const int RequestStartIndex = 0;
+    private const int CommandCountIndex = 1;
+    private const int RumbleContentsStartIndex = 2;
+    private const int CommandIndex = 10;
+    private const int ArgumentsStartIndex = 11;
+    private const int BluetoothPacketSize = 49;
+    private const int USBPacketSize = 64;
+    private const int RumbleLength = CommandIndex - RumbleContentsStartIndex;
+
+    private readonly int _packetSize;
+    private int MaxArgsLength => _packetSize - ArgumentsStartIndex;
+    private readonly int _argsLength;
+    
+    [InlineArray(USBPacketSize)]
+    private struct CommandBuffer
+    {
+        private byte _firstElement;
+    }
+
+    private readonly CommandBuffer _raw;
+    
+    public SubCommand(
+        SubCommandOperation subCommandOperation, 
+        uint commandCount, 
+        ReadOnlySpan<byte> args = default, 
+        ReadOnlySpan<byte> rumble = default,
+        bool useUSBPacketSize = false)
+    {
+        _packetSize = useUSBPacketSize ? USBPacketSize : BluetoothPacketSize;
+        
+        // Default to stopping the rumble
+        if (rumble.IsEmpty)
+        {
+            rumble = _stopRumbleBuf;
+        }
+        else if (rumble.Length != RumbleLength) // Check the rumble length if user provided
+        {
+            throw new ArgumentException($@"Rumble span is not correct size. Expected: {RumbleLength} Received: {rumble.Length}", nameof(rumble));
+        }
+
+        // Check the args length
+        if (args.Length > MaxArgsLength)
+        {
+            throw new ArgumentException($@"Args span is too large. Expected at most: {RumbleLength} Received: {rumble.Length}", nameof(args));
+        }
+        
+        _argsLength = args.Length;
+        _raw[RequestStartIndex] = 0x01; // Always
+        _raw[CommandCountIndex] = (byte)(commandCount & 0x0F); // Command index only uses 4 bits
+        _raw[CommandIndex] = (byte)subCommandOperation;
+        
+        rumble.CopyTo(_raw[RumbleContentsStartIndex..]);
+        args.CopyTo(_raw[ArgumentsStartIndex..]);
+    }
+    
+    public static implicit operator ReadOnlySpan<byte>(SubCommand subCommand) => subCommand._raw;
+    
+    public SubCommandOperation Operation => (SubCommandOperation)_raw[CommandIndex];
+
+    public ReadOnlySpan<byte> Arguments => ((ReadOnlySpan<byte>)_raw).Slice(ArgumentsStartIndex, _argsLength);
+    
+    public ReadOnlySpan<byte> Rumble => ((ReadOnlySpan<byte>)_raw).Slice(RumbleContentsStartIndex, RumbleLength);
+    
+    public override string ToString()
+    {
+        var output = new StringBuilder();
+
+        output.Append($"Subcommand {Operation:X2} sent.");
+        
+        if (_argsLength > 0)
+        {
+            output.Append(" Data:");
+
+            foreach (var arg in Arguments)
+            {
+                output.Append($" {arg:X2}");
+            }
+        }
+        
+        output.Append("Rumble: ");
+
+        if (_stopRumbleBuf.AsSpan().SequenceEqual(Rumble))
+        {
+            output.Append(" <Stop Sequence>");
+        }
+        else
+        {
+            foreach (var rumbleVal in ((ReadOnlySpan<byte>)_raw).Slice(ArgumentsStartIndex, _argsLength))
+            {
+                output.Append($" {rumbleVal:X2}");
+            }
+        }
+
+        return output.ToString();
+    }
+}
